+++ ChangeLog.md
@@ -10,6 +10,7 @@ Incoming
     * Removed `Pattern::quote()`. Use `preg::quote()`, which behaves in exactly the same way.
     * Move `Pattern::unquote()` to `preg::unquote()`, which behaves in exactly the same way.
     * Removed `pattern()->remove()->all()`. Use `pattern()->prune()` instead.
+    * Removed `pattern()->match()->asArray()`. Use `Detail.groups()` or `Detail.namedGroups()` instead.
 * Other
     * Using TDD in T-Regx, it wasn't hard to reach 100% coverage quite easily in T-Regx. In order to make the tests even
       better, we decided that the integration tests won't report any coverage, since it doesn't provide any more